app/requirements.txt
Django>=1.8,<2.0
psycopg2
gunicorn==19.3.0
djangorestframework
django-extensions